feat: inject script earlier
